Update DistPoint3Tetrahedron3.h
The point-tetrahedron distance query code failed to compile because of a mismatch in index type. (Reorganization of my internal tools caused the template instantiation tool not to launch when Tetrahedron3 was modified.

+++ GTE/Mathematics/DistPoint3Tetrahedron3.h
@@ -3,7 +3,7 @@
 // Distributed under the Boost Software License, Version 1.0.
 // https://www.boost.org/LICENSE_1_0.txt
 // https://www.geometrictools.com/License/Boost/LICENSE_1_0.txt
-// Version: 6.0.2022.01.06
+// Version: 6.0.2022.01.30
 
 #pragma once
 
@@ -69,7 +69,7 @@ namespace gte
             {
                 if (Dot(planes[i].normal, point) >= planes[i].constant)
                 {
-                    std::array<int32_t, 3> indices{ 0, 0, 0 };
+                    std::array<size_t, 3> indices{ 0, 0, 0 };
                     tetrahedron.GetFaceIndices(i, indices);
                     Triangle3<T> triangle(
                         tetrahedron.v[indices[0]],
